requirements.txt
requests>=2.28.0
pyyaml>=6.0
playwright>=1.40.0    # For WeChat Reading upload
yt-dlp>=2024.1.0      # Video download
openai-whisper>=20231117  # Local speech recognition
torch>=2.0.0          # Required by Whisper
# google-generativeai>=0.3.0  # Removed: Legacy Gemini code moved to legacy/
opencc-python-reimplemented  # Traditional to Simplified Chinese conversion
qwen-asr                     # New Qwen3 ASR engine
